Close the db module to the crate that owns it
Nothing outside runtara-environment names db any more, so it stops being pub.
Its tenant-metrics tests move in beside the query they pin — they exercise it
below the bucket cap the handler enforces, so they cannot go through the public
API — and the rest retarget onto InstanceRepository, which is the contract they
were really asserting on.

Making the module private is what finally let dead-code analysis see it: the
Instance row struct had no reader left, and five columns of the list
projection, including every instance's output blob, were being fetched on every
page and dropped. Both are gone.
